+++ lib/phone_numbers_parser.dart
@@ -3,6 +3,6 @@
 /// More dartdocs go here.
 library phone_number_parser;
 
-export 'src/phone_number_parser_base.dart';
+export 'src/phone_numbers_parser_base.dart';
 
 // TODO: Export any libraries intended for clients of this package.