+++ lib/src/parsers/light_phone_parser.dart
@@ -1,4 +1,4 @@
-import 'package:phone_numbers_parser/phone_number_parser.dart';
+import 'package:phone_numbers_parser/phone_numbers_parser.dart';
 import 'package:phone_numbers_parser/src/models/phone_number.dart';
 import 'package:phone_numbers_parser/src/models/phone_number_impl.dart';
 import 'package:phone_numbers_parser/src/models/phone_number_type.dart';
@@ -7,6 +7,9 @@ import 'package:phone_numbers_parser/src/parsers/_international_prefix_parser.da
 import 'package:phone_numbers_parser/src/parsers/_iso_code_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_national_prefix_parser.dart';
 import 'package:phone_numbers_parser/src/parsers/_validator.dart';
+import 'package:phone_numbers_parser/src/utils/_metadata_finder.dart';
+
+import '_text_parser.dart';
 
 /// light parser, faster and lighter for tree shaking
 ///