+++ lib/src/phone_number_parser_base.dart
@@ -1,2 +1,0 @@
-export 'utils/_metadata_finder.dart';
-export 'parsers/_text_parser.dart';